refactor: move pytest fixtures into single file
